Major refactor of ViewResult.
Split up `executeResult()` into smaller, more manageable functions.
Most of them are static and functional - so they'll be easier to test / reason about.
I've also cleaned up the way state is stored for whether a view or layout is supplied with a path, a direct string, or not at all.
Now it all uses the TemplateSource enum and is less confusing.