Reverse decision in a8dd4e12fbc5d73d385196afdc0163ed901126cd
See #15.

The purpose of the commit (or the pull request) was to make it easier
to use an async value with your result.

Before we could already do this:

	return getSomeFuture() >> function(data) return new JsonResult(data);

And I wanted to use `Futuristic` to make it possible to do this:

	return new JsonResult( getSomeFuture() );

But it made unit testing much more verbose. We went from:

	whenIVisit("/tags.json").onTheApp(ufApp).itShouldReturn(JsonResult, function(jr) {
		Assert.equals(0, jr.content.tags.length);
	});

To something like this:

	whenIVisit("/tags.json").onTheApp(ufApp).itShouldReturn(JsonResult, function(jr) {
		var done = Assert.createAsync();
		jr.contentFuture.handle(function(data) {
			Assert.equals(0, data.tags.length);
			done();
		});
	});

... which I thought made it too much hassle (when writing tests) and not enough benefit.

I wonder if instead I can offer this syntax:

	return getSomeFuture() >> JsonResult.create();

Where create is simply an alias for new?