Remove SysUtil from ufront-mvc.
It is used in our ufront CLI tool, but everywhere else it was fairly unnecessary.

Heading into a stable 1.0.0 release, I feel this is something I don't want to support
long term as a public API.  So I'm getting rid of it for now.  Feel free to copy it into
your own project if you were using it.