Major cleanup, refactor, documentation, fixes and testing of `ufront.web.url` package.
Notes:

- I've deleted SegmentToParamUrlFilter, not really necessary with our routing.
- I've deleted the UrlDirection enum - it was no longer used anywhere.
- UFUrlFilter no longer requires a `HttpRequest` on `filterIn` and `filterOut`. That was fairly bad design.
- Through unit testing I've come to understand exactly what it is `VirtualUrl` is supposed to be, and what `isPhysical` means. It's now documented.
- Through unit testing a number of small bugs were fixed.
- Documentation added.
- Code cleanup.

These may result in some breaking changes, but with so few people using Ufront currently, and with these classes not often being used outside of ufront-mvc, I doubt it.
It's best to get them out of the way now.

After all that, it turns out it's a pretty helpful package.  I'll have to investigate what more we can do, but some ideas:

- Have a ViewResult helper for url(), tied into Context.generateUri()
- Have a HttpContext.generateRoute() macro that takes a type safe contorller.subcontroller.action(param1,param2) string and generates a route, and passes it through `Context.generateUri()`