Fix for `ViewResult.inferLayoutFromContext` on PHP with no default layout
The dependency injector returns null for `layoutString`, and then we perform
a substring on it. Before we had a try{} catch() {} block to check if it didn't
work. But PHP didn't throw an error here, but stumbled right over it resulting in
a layout path of "/null".

This is a workaround.