Only map injection for UFViewEngine to configuration.viewEngine
Don't also have a mapping for configuration.viewEngine to itself,
even though we use this pattern elsewhere.  The reason is that
minject would require us to either have 2 singletons (one for each
mapping) or to instantiate the UFViewEngine immediately.

Because this runs before the first HttpContext is ready, we can't
get the UFViewEngine yet, so we have to either a) Only have 1 mapping
or b) Have 2 singletons.

I'm choosing to only map for `UFViewEngine`.  You can cast to another
type if you really need it.