Slightly more nuanced injection in CacheSession.
Firstly, before there was an assumption CacheConnection would
be injected before injectConfig() was run, which I don't think
is guaranteed by minject.

Secondly, now we don't need a CacheConnection property, when we
only use it once, and then keep the cache but have no need for the
connection object.

Thirdly, if the value was null (not sure why minject didn't throw an
error here), then the property access on null was a fatal error on PHP,
meaning that it bombed out rather than being logged properly.  Now we
cover ourselves for that case.