Try prevent ApiMacros and ControllerMacros from "eating" our error messages.
Previously, these macros had some try/catch blocks that caught too many errors.
In particular, if for example UserController needed UserModel, but UserModel had a syntax error, it would report an error in UserController, not UserModel.

It turns out this is quite difficult to correct:

- You can use Lib.rethrow() to throw the original error at the original location
- Sometimes "the original location" is ApiMacros.hx, and so it isn't helpful.
- My current workaround is not ideal, but it is to give a warning at one position, and re-throw an error at another.
- This results in one correct position plus one or more bogus positions, but it's better than resulting in only one position that might not correct.

I'll keep looking for a better solution.

Fixes #35