Refactor ViewResult so that it has `renderResult():Surprise<String,Error>`
This way it can be used outside of a Http environment.
For example, if you want to render a page for offline viewing, or render a HTML email etc.

Please note, if you have a sub-class of ViewResult that overrides
`writeResponse()`, you'll need to change the signiature to no longer
have the `combinedData:TemplateData` sent through.

See #6