Add erazorHtml templating engine that escapes HTML by default.
If you don't specify your templating engines in your UfrontConfiguration,
this will now be the default before normal `erazor`.

If you want the unsafe one, specify it in your `UfrontConfiguration`.

In a future version I'll probably deprecate the unsafe one, or at least rename it to make it more obvious it is unsafe.