Add MockHttpRequest, MockHttpResponse, remove need for mockatoo in TestUtils
Mockatoo is no longer needed to run testutils.

The reasoning here is that it's one extra tool, (and quite an opaque/hard-to-learn one), that we don't need for such a simple use case.

If someone still wants to use it in their project they can, but we won't require it if all they want is to mock a HttpContext.