Fix ufront.api.ApiMacros with tink-core 1.18+
Since
https://github.com/haxetink/tink_core/commit/f85e6a090a94158868ff1ce110be8b2e83904a08
T can be cast to tink.core.Future<T>, and therefore unifies with
tink.core.Future<Dynamic>. This means that we cannot check whether the
return types unify with Future/Surprise, but instead we have to check
the type with the similar code to how we retrieve the type params.

This solves the error:

`Expected type String to be a tink.core.Future`

This was happening during the unit tests because String unifies with
tink.core.Future, so String was assumed to be a Future. This caused an
error when trying to get the supposed "Future's" type parameters.